Kill warnings in TIING_GUARD code that we may soon delete
git-svn-id: http://svn.research.cc.gatech.edu/kaos/evpath/trunk@26775 8ba6a8b9-845b-0410-94f0-974be2f5df47

+++ cmsockets.c
@@ -124,7 +124,7 @@ static atom_t CM_IP_ADDR = -1;
 
 static int do_timing = 0;
 #define TIMING_GUARD_START {     struct timeval t0,t1,diff; gettimeofday(&t0, NULL);
-#define TIMING_GUARD_STOP gettimeofday(&t1, NULL);    timersub(&t1, &t0, &diff); if (diff.tv_sec > 0) if(do_timing)fprintf(stderr, "TIME GUARD at %s:%d exceeded, time was was <%ld.%06d> secs\n", __FILE__, __LINE__, diff.tv_sec, diff.tv_usec);}
+#define TIMING_GUARD_STOP gettimeofday(&t1, NULL);    timersub(&t1, &t0, &diff); if (diff.tv_sec > 0) if(do_timing)fprintf(stderr, "TIME GUARD at %s:%d exceeded, time was was <%ld.%06ld> secs\n", __FILE__, __LINE__, (long)diff.tv_sec, (long)diff.tv_usec);}
 
 static int
 check_host(hostname, sin_addr)
@@ -1075,7 +1075,7 @@ attr_list attrs;
 		    strftime(tmbuf, sizeof tmbuf, "%Y-%m-%d %H:%M:%S", nowtm);
 		    snprintf(stop_buf, sizeof(stop_buf), "%s.%06ld", tmbuf, (long)stop.tv_usec);
 
-		    fprintf(stderr, "TIME GUARD in writev() at %s:%d exceeded, time was was <%ld.%06d> secs\n", __FILE__, __LINE__, diff.tv_sec, diff.tv_usec);
+		    fprintf(stderr, "TIME GUARD in writev() at %s:%d exceeded, time was was <%ld.%06ld> secs\n", __FILE__, __LINE__, (long)diff.tv_sec, (long)diff.tv_usec);
 		    fprintf(stderr, "Original write length was %d bytes, start time was %s, end was %s\n", orig_len, start_buf, stop_buf);
 		}
 	    }
